4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,16 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -22,11 +15,7 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(request.user.id)
 
 
   const blog = new Blog({
@@ -46,6 +35,14 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const userid = request.user.id
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (userid.toString()!== blog.user.toString()){
+    return response.status(401).json({ error: 'wrong user' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
